+++ src/index.ts
@@ -1,9 +1,4 @@
-export {
-  runWxmlPipeline,
-  type RunWxmlPipelineOptions,
-  WXML_FORMAT_ON_ERROR,
-  type WxmlFormatOnError,
-} from './pipeline/run-wxml-pipeline'
+export { runWxmlPipeline, type RunWxmlPipelineOptions } from './pipeline/run-wxml-pipeline'
 
 export {
   defaultExport as default,
@@ -14,8 +9,7 @@ export {
   printers,
 } from './plugin'
 
-export { WXML_REPORT_LEVEL } from './plugin-options'
-export type { WxmlPluginOptions, WxmlReportLevel } from './plugin-options'
+export type { WxmlPluginOptions } from './plugin-options'
 
 export { resolveSelfCloseExcludeSet, selfCloseTags } from './self-close-tags'
 export type { WxmlSelfCloseExclude } from './self-close-tags'